voice-agent/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Audio processing
pydub==0.25.1

# Scientific computing
numpy==1.24.3

# Speech recognition (install last)
faster-whisper==1.0.0
